finished maps and stages

+++ Adventure Island/Assets/Scripts/SC_FireObstacle.cs
@@ -4,12 +4,20 @@ public class SC_FireObstacle : MonoBehaviour
 {
     private void OnCollisionEnter2D(Collision2D collision)
     {
-       
         if (collision.gameObject.CompareTag("Player"))
         {
             if (!SC_PlayerController.instance.GetImmune())
             {
-                SC_PlayerController.instance.kill();
+                if (SC_PlayerController.instance.GetMount() != null)
+                {
+                    SC_PlayerController.instance.GetMount().Dismount();
+                    SC_PlayerController.instance.SetMount(null);
+                    gameObject.SetActive(false);
+                }
+                else
+                {
+                    SC_PlayerController.instance.kill();
+                }
             }
             else
             {
